backend/requirements.txt
# PyTorch and Vision
torch>=2.5.0
torchvision>=0.20.0

# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.12.5

# Image Processing
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
scikit-image>=0.22.0

# Training Utilities
tqdm>=4.66.0
matplotlib>=3.8.0

# HTTP Client
httpx>=0.24,<0.26
